requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django_debug_toolbar==3.8.1
idna==3.4
Pillow==9.4.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
